+++ listers/static.go
